Fix patching of Sync/Async client to enable auth
