+++ diracx-client/src/diracx/client/patches/client/sync.py
@@ -1,9 +1,4 @@
-"""Patches for the autorest-generated jobs client.
-
-This file can be used to customize the generated code for the jobs client.
-When adding new classes to this file, make sure to also add them to the
-__all__ list in the corresponding file in the patches directory.
-"""
+"""Patches for the autorest-generated client to enable authentication."""
 
 from __future__ import annotations
 
@@ -11,8 +6,140 @@ __all__ = [
     "Dirac",
 ]
 
+from importlib.metadata import PackageNotFoundError, distribution
+from pathlib import Path
+from typing import Any, Optional
+
+from azure.core.credentials import AccessToken, TokenCredential
+from azure.core.pipeline import PipelineRequest
+from azure.core.pipeline.policies import BearerTokenCredentialPolicy
+from diracx.core.preferences import DiracxPreferences, get_diracx_preferences
+
+from .common import get_openid_configuration, get_token
 from ..._generated._client import Dirac as _Dirac
-from .common import DiracAuthMixin
 
 
-class Dirac(DiracAuthMixin, _Dirac): ...
+class SyncDiracTokenCredential(TokenCredential):
+    """Tailor get_token() for our context"""
+
+    def __init__(
+        self,
+        location: Path,
+        token_endpoint: str,
+        client_id: str,
+        *,
+        verify: bool | str = True,
+    ) -> None:
+        self.location = location
+        self.verify = verify
+        self.token_endpoint = token_endpoint
+        self.client_id = client_id
+
+    def get_token(
+        self,
+        *scopes: str,
+        claims: Optional[str] = None,
+        tenant_id: Optional[str] = None,
+        **kwargs: Any,
+    ) -> AccessToken:
+        return get_token(
+            self.location,
+            kwargs.get("token"),
+            self.token_endpoint,
+            self.client_id,
+            self.verify,
+        )
+
+
+class SyncDiracBearerTokenCredentialPolicy(BearerTokenCredentialPolicy):
+    """Custom BearerTokenCredentialPolicy tailored for our use case.
+
+    * It does not ensure the connection is done through https.
+    * It does not ensure that an access token is available.
+    """
+
+    # Make mypy happy
+    _token: Optional[AccessToken] = None
+
+    def __init__(
+        self, credential: SyncDiracTokenCredential, *scopes: str, **kwargs: Any
+    ) -> None:
+        super().__init__(credential, *scopes, **kwargs)
+
+    def on_request(self, request: PipelineRequest) -> None:
+        """Authorization Bearer is optional here.
+        :param request: The pipeline request object to be modified.
+        :type request: ~azure.core.pipeline.PipelineRequest
+        :raises: :class:`~azure.core.exceptions.ServiceRequestError`
+        """
+        if not isinstance(self._credential, TokenCredential):
+            raise NotImplementedError(
+                "SyncDiracBearerTokenCredentialPolicy only supports TokenCredential"
+            )
+
+        self._token = self._credential.get_token("", token=self._token)
+        if not self._token.token:
+            # If we are here, it means the token is not available
+            # we suppose it is not needed to perform the request
+            return
+
+        self._update_headers(request.http_request.headers, self._token.token)
+
+
+class Dirac(_Dirac):
+    """This class inherits from the generated Dirac client and adds support for tokens,
+    so that the caller does not need to configure it by itself.
+    """
+
+    def __init__(
+        self,
+        endpoint: str | None = None,
+        client_id: str | None = None,
+        diracx_preferences: DiracxPreferences | None = None,
+        verify: bool | str = True,
+        **kwargs: Any,
+    ) -> None:
+        diracx_preferences = diracx_preferences or get_diracx_preferences()
+        self._endpoint = str(endpoint or diracx_preferences.url)
+        if verify is True and diracx_preferences.ca_path:
+            verify = str(diracx_preferences.ca_path)
+        kwargs["connection_verify"] = verify
+        self._client_id = client_id or "myDIRACClientID"
+
+        # Get .well-known configuration
+        openid_configuration = get_openid_configuration(self._endpoint, verify=verify)
+
+        try:
+            self.client_version = distribution("diracx").version
+        except PackageNotFoundError:
+            try:
+                self.client_version = distribution("diracx-client").version
+            except PackageNotFoundError:
+                print("Error while getting client version")
+                self.client_version = "Unknown"
+
+        # Setting default headers
+        kwargs.setdefault("base_headers", {})[
+            "DiracX-Client-Version"
+        ] = self.client_version
+
+        # Initialize Dirac with a Dirac-specific token credential policy
+        # We need to ignore types here because mypy complains that we give
+        # too many arguments to "object" constructor as this is a mixin
+
+        super().__init__(  # type: ignore
+            endpoint=self._endpoint,
+            authentication_policy=SyncDiracBearerTokenCredentialPolicy(
+                SyncDiracTokenCredential(
+                    location=diracx_preferences.credentials_path,
+                    token_endpoint=openid_configuration["token_endpoint"],
+                    client_id=self._client_id,
+                    verify=verify,
+                ),
+            ),
+            **kwargs,
+        )
+
+    @property
+    def client_id(self):
+        return self._client_id
